fix: respects Dependency Inversion principle by extending XYZ with OSM
