+++ src/serializer/MapFishPrintV3StamenSerializer.ts
@@ -1,9 +1,9 @@
 import OlLayer from 'ol/layer/Layer';
 import OlSource from 'ol/source/Source';
 import OlSourceStamen from 'ol/source/Stamen';
-import MapFishPrintV3OSMSerializer from './MapFishPrintV3OSMSerializer';
+import MapFishPrintV3XYZSerializer from './MapFishPrintV3XYZSerializer';
 
-export class MapFishPrintV3StamenSerializer extends MapFishPrintV3OSMSerializer {
+export class MapFishPrintV3StamenSerializer extends MapFishPrintV3XYZSerializer {
 
   validateSource(source: OlSource): source is OlSourceStamen {
     return source instanceof OlSourceStamen;
